% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{add_arma}
\alias{add_arma}
\alias{add_arma.issm.component}
\title{ARMA Component}
\usage{
add_arma(x, ...)

\method{add_arma}{issm.component}(x, order = c(0, 0), ar = 0, ma = 0, mu = 0, ...)
}
\arguments{
\item{x}{an object of class issm.component or other supported class.}

\item{...}{additional parameters.}

\item{order}{the ar and ma orders.}

\item{ar}{a vector of ar coefficients.}

\item{ma}{a vector of ma coefficients.}

\item{mu}{the mean parameter (defaults to zero) of the ARMA process.}
}
\value{
An object of class issm.component updated with the ARMA
component.
}
\description{
ARMA Component
}
