% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{add_custom}
\alias{add_custom}
\alias{add_custom.issm.component}
\title{Custom Component}
\usage{
add_custom(x, ...)

\method{add_custom}{issm.component}(x, custom = NULL, ...)
}
\arguments{
\item{x}{an object of class issm.component or other supported class.}

\item{...}{additional parameters.}

\item{custom}{a matrix of custom components}
}
\value{
An object of class issm.component updated with the custom
components.
}
\description{
Custom Component
}
