% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{add_polynomial}
\alias{add_polynomial}
\alias{add_polynomial.issm.component}
\title{Polynomial Trend Component}
\usage{
add_polynomial(x, ...)

\method{add_polynomial}{issm.component}(
  x,
  order = 1,
  alpha = 0.1,
  beta = 0.01,
  phi = 1,
  l0 = 100,
  b0 = 1,
  ...
)
}
\arguments{
\item{x}{an object of class issm.component or other supported class.}

\item{...}{additional parameters.}

\item{order}{the order of the polynomial (min 1 and max 2).}

\item{alpha}{the decay coefficient on the error of the level.}

\item{beta}{the decay coefficient on the error of the slope.}

\item{phi}{dampening parameter for the slope.}

\item{l0}{initial level.}

\item{b0}{initial slope.}
}
\value{
An object of class issm.component updated with the polynomial trend
component.
}
\description{
Polynomial Trend Component
}
