% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{add_transform}
\alias{add_transform}
\alias{add_transform.tsissm.component}
\title{Transform}
\usage{
add_transform(x, ...)

\method{add_transform}{tsissm.component}(x, method = "box-cox", lambda = 1, lower = 0, upper = 1, ...)
}
\arguments{
\item{x}{an object of class tsissm.component or other supported class.}

\item{...}{additional parameters.}

\item{method}{a valid transform.}

\item{lambda}{the Box-Cox parameter.}

\item{lower}{the lower bound for the transform.}

\item{upper}{the upper bound for the transform.}
}
\value{
An object of class tsissm.component updated with the transformation.
}
\description{
Transform
}
\details{
The inverse transform is applied to the simulated series. Valid methods
are the \dQuote{box-cox}, \dQuote{logit}, \dQuote{softplus-logit} and \dQuote{sigmoid}
transforms.
}
