\name{calendar_eom}
\alias{calendar_eom}
\title{
End of Month Date
}
\description{
Returns the last day of the month from a Date within the month.
}
\usage{
calendar_eom(date)
}
\arguments{
\item{date}{
A Date vector
}
}
\details{
Given a Date (such as 2019-01-02), will return the last Date within that year month.
}
\value{
Date object
}
\author{
Alexios Galanos
}
