% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{calendar_eow}
\alias{calendar_eow}
\title{End of Week Date}
\usage{
calendar_eow(date, day = 7, ...)
}
\arguments{
\item{date}{a Date vector}

\item{day}{a value between 1 (Monday) and 7 (Sunday).}

\item{\dots}{not used}
}
\value{
Date object
}
\description{
Returns the last day of the week from a Date given a choice of week days.
}
\details{
Given a Date (such as 2019-01-02) and a day of 7, will return the Date for
the Sunday at or immediately after that. The week starting day is Monday
(1). A simple use case is when one wants to aggregate daily data to a
regular weekly sequence.
}
\author{
Alexios Galanos
}
