\name{calendar_eoy}
\alias{calendar_eoy}
\title{
End of Year Date
}
\description{
Returns the last day of the year from a Date.
}
\usage{
calendar_eoy(date)
}
\arguments{
\item{date}{
A Date vector
}
}
\details{
Given a date (such as 2019-01-02), will return the last date within that year.
}
\value{
Date object
}
\author{
Alexios Galanos
}
