% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscutils.R
\name{check_xreg}
\alias{check_xreg}
\alias{check_newxreg}
\title{Checks on regressor matrix.}
\usage{
check_xreg(xreg, valid_index)

check_newxreg(newdata, xnames, h = 1, forc_dates = NULL)
}
\arguments{
\item{xreg}{an xts matrix of named regressors.}

\item{valid_index}{a vector of dates against which the xreg matrix index is
compared for validity.}

\item{newdata}{an xts matrix of out of named sample regressors.}

\item{xnames}{names of regressors used in sample.}

\item{h}{the forecast horizon}

\item{forc_dates}{an optional vector of forecast dates. This is used if
newdata is not an xts matrix in which case it formats the data into such
using the forc_dates vector.}
}
\value{
Returns the xts input matrix if checks are passed else raises an
error.
}
\description{
Used internally by other packages, these functions provides some commonly
used validation checks on regressor matrices in both in and out of sample.
}
