% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{lines.issm.component}
\alias{lines.issm.component}
\alias{lines}
\title{Add Connected Line Segments to a Simulation Object}
\usage{
\method{lines}{issm.component}(x, y = NULL, type = "l", ...)
}
\arguments{
\item{x}{an object of class issm.component or other supported class.}

\item{y}{not used.}

\item{type}{character indicating the type of plotting.}

\item{...}{additional parameters passed to the lines function.}
}
\value{
a line plot.
}
\description{
Add Connected Line Segments to a Simulation Object
}
\details{
Overlays the simulated series from the object (x), and is meant to be
used when plotting different simulations from the same series
for comparison.
}
