% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{logit}
\alias{logit}
\title{The logit transformation}
\usage{
logit(lower = 0, upper = 1, ...)
}
\arguments{
\item{lower}{lower bound of the variable.}

\item{upper}{upper bound of the variable.}

\item{...}{not currently used.}
}
\value{
A list with the transform and inverse functions.
}
\description{
The logit transformation as an alternative to the Box Cox for bounded
outcomes.
}
\author{
Alexios Galanos
}
