% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{logit_transform}
\alias{logit_transform}
\alias{logit}
\alias{logit_inverse}
\title{The logit transformation}
\usage{
logit_transform(x, lower = 0, upper = 1, ...)
}
\arguments{
\item{x}{a numeric vector (or object which can be coerced to such).}

\item{lower}{lower bound of the variable.}

\item{upper}{upper bound of the variable.}

\item{...}{not currently used.}
}
\value{
The transformed value or list of transformation functions.
}
\description{
The logit transformation as an alternative to the Box Cox for bounded
outcomes.
}
\details{
The logit function without a data argument returns a list of the 2 functions
(transform and inverse) with hard coded min and max bounds to be used in
other functions.
}
\author{
Alexios Galanos
}
