% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{plot.issm.component}
\alias{plot.issm.component}
\alias{plot}
\title{Plot Simulation Object}
\usage{
\method{plot}{issm.component}(x, y = c("simulated", "components"), ...)
}
\arguments{
\item{x}{an object of class issm.component or other supported class.}

\item{y}{the type of output to plot.}

\item{...}{additional parameters passed to the \code{\link[zoo]{plot.zoo}} function.}
}
\value{
a plot of the simulated series or multiple plots of the simulation components.
}
\description{
Plot Simulation Object
}
