% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonal_dummies}
\alias{seasonal_dummies}
\title{Seasonal Dummies}
\usage{
seasonal_dummies(y = NULL, n = nrow(y), seasons = 12)
}
\arguments{
\item{y}{optional data series.}

\item{n}{if y is missing, then the length of the series is required.}

\item{seasons}{number of seasons in a cycle.}
}
\value{
Either a matrix (if y is missing or y is not an xts vector) or an
xts matrix (when y is an xts vector).
}
\description{
Creates a matrix of seasonal dummies.
}
\details{
Generates seasons-1 dummy variables.
}
\examples{

head(seasonal_dummies(n=100, seasons=12))

}
\author{
Alexios Galanos
}
