% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{tsdecompose.tsissm.component}
\alias{tsdecompose.tsissm.component}
\alias{tsdecompose}
\title{State Decomposition}
\usage{
\method{tsdecompose}{tsissm.component}(object, ...)
}
\arguments{
\item{object}{an object of class tsissm.component or other supported class.}

\item{...}{additional parameters.}
}
\value{
A matrix of the simplified state decomposition.
}
\description{
State Decomposition
}
\details{
Creates a simplified decomposition of the states and aligns their
time indices so that the sum up to the simulated component per period.
}
