% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsaux-package.R
\docType{package}
\name{tsaux-package}
\alias{tsaux}
\alias{tsaux-package}
\title{tsaux: Time Series Forecasting Auxilliary Functions}
\description{
General purpose calendar and auxiliary functions including normalizing transformations, outlier detection and forecast metrics for time series.
}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com} (\href{https://orcid.org/0009-0000-9308-0457}{ORCID}) [copyright holder]

}
\keyword{internal}
