% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tslinear.R
\name{tslinear}
\alias{tslinear}
\title{Linear Time Series Filter}
\usage{
tslinear(y, trend = FALSE, seasonal = FALSE, xreg = NULL, frequency = 1, ...)
}
\arguments{
\item{y}{a vector.}

\item{trend}{whether to include a linear trend.}

\item{seasonal}{whether to include seasonal dummies.}

\item{xreg}{an optional matrix of regressors.}

\item{frequency}{the frequency of the series (required if seasonal is TRUE).}

\item{\dots}{not currently used.}
}
\value{
An object of class \dQuote{tslinear} which also inherits
\dQuote{lm}.
}
\description{
Provides a quick linear time series filter with trend and seasonality.
}
\details{
Estimates a simple linear time series model with trend, seasonal and
regressors.
}
\author{
Alexios Galanos
}
