% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{add_anomaly}
\alias{add_anomaly}
\alias{add_anomaly.tsissm.component}
\title{Anomaly Component}
\usage{
add_anomaly(x, ...)

\method{add_anomaly}{tsissm.component}(x, time = NULL, delta = 0, ratio = 0.5, ...)
}
\arguments{
\item{x}{an object of class tsissm.component or other supported class.}

\item{...}{additional parameters.}

\item{time}{the numeric index of when the anomaly occurs. If NULL, a ranom
time will be chosen.}

\item{delta}{the autoregressive component determining the type of anomaly. A
value of zero results in an additive outlier, a value of 1 in a level shift
and anything in between a temporary change with a half life of -log(2)/log(delta).}

\item{ratio}{the anomaly to series ratio at the time it occurs. For instance, a
value of 1 means that the anomaly will jump by 100 percent compared to the
data series.}
}
\value{
An object of class tsissm.component updated with the anomaly
component.
}
\description{
Anomaly Component
}
