% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{calendar_eom}
\alias{calendar_eom}
\title{End of Month Date}
\usage{
calendar_eom(date, ...)
}
\arguments{
\item{date}{a Date vector}

\item{\dots}{not used}
}
\value{
Date object
}
\description{
Returns the last day of the month from a Date within the month.
}
\details{
Given a Date (such as 2019-01-02), will return the last Date within that
year month.
}
\author{
Alexios Galanos
}
