\name{calendar_eoq}
\alias{calendar_eoq}
\title{
End of Quarter Date
}
\description{
Returns the last day of the quarter from a Date.
}
\usage{
calendar_eoq(date)
}
\arguments{
\item{date}{
A Date vector
}
}
\details{
Given a date (such as 2019-01-02), will return the last date within that year quarter.
}
\value{
Date object
}
\author{
Alexios Galanos
}
