% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{calendar_eoy}
\alias{calendar_eoy}
\title{End of Year Date}
\usage{
calendar_eoy(date, ...)
}
\arguments{
\item{date}{a Date vector}

\item{\dots}{not used}
}
\value{
Date object
}
\description{
Returns the last day of the year from a Date.
}
\details{
Given a date (such as 2019-01-02), will return the last date within that
year.
}
\author{
Alexios Galanos
}
