% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{future_dates}
\alias{future_dates}
\title{Generate Regular Interval Future Dates}
\usage{
future_dates(start, frequency, n = 1)
}
\arguments{
\item{start}{a Date string for the start date.}

\item{frequency}{frequency of the interval (daily, weekly, monthly or
yearly).}

\item{n}{number of future periods to generate dates for.}
}
\value{
A Date vector
}
\description{
Generates regular interval future dates for use in forecast routine.
}
\author{
Alexios Galanos
}
