% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{lines.tsissm.component}
\alias{lines.tsissm.component}
\alias{lines}
\title{Add Connected Line Segments to a Simulation Object}
\usage{
\method{lines}{tsissm.component}(x, y = NULL, type = "l", ...)
}
\arguments{
\item{x}{an object of class tsissm.component or other supported class.}

\item{y}{not used.}

\item{type}{character indicating the type of plotting.}

\item{...}{additional parameters passed to the lines function.}
}
\description{
Add Connected Line Segments to a Simulation Object
}
