% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{plot.tsissm.component}
\alias{plot.tsissm.component}
\alias{plot}
\title{Plot Simulation Object}
\usage{
\method{plot}{tsissm.component}(x, y = c("simulated", "components"), ...)
}
\arguments{
\item{x}{an object of class tsissm.component or other supported class.}

\item{y}{the type of output to plot.}

\item{...}{additional parameters passed to the \code{\link{plot.zoo}} function.}
}
\description{
Plot Simulation Object
}
