% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal.R
\name{seasonality_test}
\alias{seasonality_test}
\title{Simple Seasonality Test}
\usage{
seasonality_test(x, frequency = NULL)
}
\arguments{
\item{x}{an (xts) vector (usually of a stationary series).}

\item{frequency}{overrides any frequency automatically identified in the
index of x.}
}
\value{
Logical.
}
\description{
Checks for the presence of seasonality based on the QS test of Gomez and
Maravall (1996).
}
\details{
Given the identified frequency of the xts vector (using the
\code{\link{sampling_frequency}}), the function checks for seasonality at
that frequency. The frequency can be overridden by directly supplying a
frequency argument, in which case y does not need to be a xts vector.
}
\references{
\insertRef{Gomez1995}{tsaux}
}
\author{
Alexios Galanos
}
