% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsextra-package.R
\docType{package}
\name{tsaux-package}
\alias{tsaux}
\alias{tsaux-package}
\title{tsaux: Time Series Forecasting Auxiliary Functions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

General purpose calendar and auxiliary functions including normalizing transformations, outlier detection and forecast metrics for time series.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tsmodels/tsaux}
  \item Report bugs at \url{https://github.com/tsmodels/tsaux/issues}
}

}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com} (\href{https://orcid.org/0009-0000-9308-0457}{ORCID}) [copyright holder]

}
\keyword{internal}
