% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsaux-package.R
\docType{package}
\name{tsaux-package}
\alias{tsaux}
\alias{tsaux-package}
\title{tsaux: Auxilliary Functions for tsmodels framework}
\description{
General purpose calendar and auxiliary functions.
}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com}

Other contributors:
\itemize{
  \item James Nesbit \email{jmcgnesbit@gmail.com} [contributor]
  \item Keith O'Hara \email{kth.ohr@gmail.com} [contributor]
}

}
\keyword{internal}
