% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tstransform}
\alias{tstransform}
\title{General transformation function}
\usage{
tstransform(method = "box-cox", lambda = NULL, lower = 0, upper = 1, ...)
}
\arguments{
\item{method}{valid methods are currently \dQuote{box-cox} and
\dQuote{logit}.}

\item{lambda}{parameter in the Box Cox transformation.}

\item{lower}{lower bound for the transformations.}

\item{upper}{upper bound for the transformations.}

\item{\dots}{additional arguments taken by the transformations.}
}
\value{
A list of functions
}
\description{
Includes the Box Cox and logit transforms. Returns a function for transform
and inverse.
}
\details{
Returns a list with functions for transform and its inverse.
}
\author{
Alexios Galanos
}
